APARTADO 2 - Práctica 3 ----------------------- Realización parcial del apartado 2 de la práctica.
-Se ha añadido una nueva carpeta (data) dentro de resources la cuál contiene los archivos .csv necesarios para la lectura de datos.
-Se han modificado algunos tipos de atributos de las tablas.
-Se han añadido las anotaciones @Column en los atributos que no son clave primaria de las clases
-Se han añadido a ProccesData.java los métodos processDimLugar() y proccesDimTiempo() el cuál lee los datos de los ficheros .csv correspondientes y los carga en la base de datos.

NOTA: la aplicación funciona correctamente hasta este punto del desarrollo.

+++ src/main/java/prac3/repositorios/RepositorioHospital.java
@@ -10,8 +10,8 @@ import java.util.List;
 public interface RepositorioHospital extends CrudRepository<DimHospital, String> {
 
     List<DimHospital> findByNombre(String nombre);
-    List<DimHospital> findByCpostal(short cpostal);
+    List<DimHospital> findByCpostal(int cpostal);
     List<DimHospital> findByAutopista(String autopista);
     List<DimHospital> findByGestor(String gestor);
-    DimHospital findByNombreAndCpostal(String nombre, short cpostal);
+    DimHospital findByNombreAndCpostal(String nombre, int cpostal);
 }
